dashboard/requirements.txt
streamlit~=1.33.0
altair~=5.3.0
pandas~=2.2.2
psycopg2~=2.9.9
python-dotenv~=1.0.1
boto3